+++ ui/views/controls/menu/menu_model_adapter_unittest.cc
@@ -66,7 +66,7 @@ class MenuModelBase : public ui::MenuModel {
     return false;
   }
 
-  bool IsItemCheckedAt(int index) const override { return false; }
+  bool IsItemCheckedAt(size_t index) const override { return false; }
 
   int GetGroupIdAt(int index) const override { return 0; }
